Crate sc_network_test

Structs

BlockImportAdapter

Implements BlockImport for any Transaction. Internally the transaction is “converted”, aka the field is set to None.

ExtrinsicBuilder

Generates Extrinsic

ForceFinalized

FullPeerConfig

Configuration for a full peer.

JustificationTestNet

PassThroughVerifier

A Verifier that accepts all blocks and passes them on with the configured finality to be imported.

Peer

PeersClient

TestNet

Transfer

Transfer used in test substrate pallet. Extrinsic is created and signed using this data.

Traits

BlockImportAdapterFull

TestClientBuilderExt

A test-runtime extensions to TestClientBuilder.

TestNetFactory

Type Definitions

Block

A test block.

Hash

A simple hash type for all our hashing.

Header

A test block’s header.

PeersFullClient

TestClient

A test client with default backend.

TestClientBuilder

A TestClient with test-runtime builder.