Struct sc_network_test::Peer

pub struct Peer<D, BlockImport> {
    pub data: D,
    // some fields omitted
}

Fields
data: D
Implementations

impl<D, B> Peer<D, B> where
B: BlockImport<Block, Error = ConsensusError> + Send + Sync,
B::Transaction: Send,

pub fn id(&self) -> PeerId

Get this peer ID.

pub fn is_major_syncing(&self) -> bool

Returns true if we’re major syncing.

pub fn select_chain(&self) -> Option<LongestChain<Backend, Block>>

pub fn num_peers(&self) -> usize

Returns the number of peers we’re connected to.

pub fn num_downloaded_blocks(&self) -> usize

Returns the number of downloaded blocks.

pub fn is_offline(&self) -> bool

Returns true if we have no peer.

pub fn request_justification(
&self,
hash: &<Block as BlockT>::Hash,
number: NumberFor<Block>
)

Request a justification for the given block.

pub fn announce_block(
&self,
hash: <Block as BlockT>::Hash,
data: Option<Vec<u8>>
)

Announces an important block on the network.

pub fn set_sync_fork_request(
&self,
peers: Vec<PeerId>,
hash: <Block as BlockT>::Hash,
number: NumberFor<Block>
)

Request explicit fork sync.

pub fn generate_blocks<F>(
&mut self,
count: usize,
origin: BlockOrigin,
edit_block: F
) -> H256 where
F: FnMut(BlockBuilder<'_, Block, PeersFullClient, Backend>) -> Block,

Add blocks to the peer – edit the block before adding

pub fn push_blocks(&mut self, count: usize, with_tx: bool) -> H256

Push blocks to the peer (simplified: with or without a TX)

pub fn push_headers(&mut self, count: usize) -> H256

Push blocks to the peer (simplified: with or without a TX)

pub fn push_blocks_at(
&mut self,
at: BlockId<Block>,
count: usize,
with_tx: bool
) -> H256

Push blocks to the peer (simplified: with or without a TX) starting from given hash.

pub fn push_blocks_at_without_informing_sync(
&mut self,
at: BlockId<Block>,
count: usize,
with_tx: bool
) -> H256

Push blocks to the peer (simplified: with or without a TX) starting from given hash without informing the sync protocol about the new best block.

pub fn push_blocks_at_without_announcing(
&mut self,
at: BlockId<Block>,
count: usize,
with_tx: bool
) -> H256

Push blocks to the peer (simplified: with or without a TX) starting from given hash without announcing the block.

pub fn push_authorities_change_block(
&mut self,
new_authorities: Vec<AuthorityId>
) -> H256

pub fn client(&self) -> &PeersClient

Get a reference to the client.

pub fn network_service(
&self
) -> &Arc<NetworkService<Block, <Block as BlockT>::Hash>>

Get a reference to the network service.

pub fn network(&self) -> &NetworkWorker<Block, <Block as BlockT>::Hash>

Get a reference to the network worker.

pub fn blockchain_canon_equals(&self, other: &Self) -> bool

Test helper to compare the blockchain state of multiple (networked) clients.

pub fn blocks_count(&self) -> u64

Count the total number of imported blocks.

pub fn failed_verifications(&self) -> HashMap<<Block as BlockT>::Hash, String>

Return a collection of block hashes that failed verification

pub fn has_block(&self, hash: &H256) -> bool