Crate sc_network_test

Structs

EmptyTransactionPool

Dummy implementation of the TransactionPool trait for a transaction pool that is always empty and discards all incoming transactions.

ForceFinalized

FullPeerConfig

Configuration for a full peer.

JustificationTestNet

PassThroughVerifier

A Verifier that accepts all blocks and passes them on with the configured finality to be imported.

Peer

TestNet

Transfer

Calls in transactions.

Enums

BlockImportAdapter

Implements BlockImport for any Transaction. Internally the transaction is “converted”, aka the field is set to None.

Extrinsic

Extrinsic for test-runtime.

PeersClient

Traits

TestClientBuilderExt

A test-runtime extensions to TestClientBuilder.

TestNetFactory

Type Definitions

Block

A test block.

Hash

A simple hash type for all our hashing.

PeersFullClient

PeersLightClient

TestClient

A test client with default backend.

TestClientBuilder

A TestClient with test-runtime builder.