Crate sc_network_test

Structs

BlockImportAdapter

Implements BlockImport for any Transaction. Internally the transaction is “converted”, aka the field is set to None.

EmptyTransactionPool

Dummy implementation of the TransactionPool trait for a transaction pool that is always empty and discards all incoming transactions.

ForceFinalized

FullPeerConfig

Configuration for a full peer.

JustificationTestNet

PassThroughVerifier

A Verifier that accepts all blocks and passes them on with the configured finality to be imported.

Peer

PeersClient

ProtocolId

Name of a protocol, transmitted on the wire. Should be unique for each chain. Always UTF-8.

TestNet

Transfer

Calls in transactions.

Enums

Extrinsic

Extrinsic for test-runtime.

Traits

BlockImportAdapterFull

TestClientBuilderExt

A test-runtime extensions to TestClientBuilder.

TestNetFactory

Type Definitions

Block

A test block.

Hash

A simple hash type for all our hashing.

PeersFullClient

TestClient

A test client with default backend.

TestClientBuilder

A TestClient with test-runtime builder.