Trait sc_network_test::TestNetFactory

pub trait TestNetFactory: Sized {
    type Specialization: NetworkSpecialization<Block> + SpecializationFactory;
    type Verifier: 'static + Verifier<Block>;
    type PeerData: Default;
    fn from_config(config: &ProtocolConfig) -> Self;

    fn make_verifier(
        &self, 
        client: PeersClient, 
        config: &ProtocolConfig, 
        peer_data: &Self::PeerData
    ) -> Self::Verifier;

    fn peer(
        &mut self, 
        i: usize
    ) -> &mut Peer<Self::PeerData, Self::Specialization>;

    fn peers(&self) -> &Vec<Peer<Self::PeerData, Self::Specialization>>;

    fn mut_peers<F: FnOnce(&mut Vec<Peer<Self::PeerData, Self::Specialization>>)>(
        &mut self, 
        closure: F
    );

    fn make_block_import(
        &self, 
        client: PeersClient
    ) -> (BoxBlockImport<Block>, Option<BoxJustificationImport<Block>>, Option<BoxFinalityProofImport<Block>>, Option<BoxFinalityProofRequestBuilder<Block>>, Self::PeerData) { ... }

    fn make_finality_proof_provider(
        &self, 
        _client: PeersClient
    ) -> Option<Arc<dyn FinalityProofProvider<Block>>> { ... }

    fn default_config() -> ProtocolConfig { ... }

    fn new(n: usize) -> Self { ... }

    fn add_full_peer(&mut self, config: &ProtocolConfig) { ... }

    fn add_full_peer_with_states(
        &mut self, 
        config: &ProtocolConfig, 
        keep_blocks: Option<u32>
    ) { ... }

    fn add_light_peer(&mut self, config: &ProtocolConfig) { ... }

    fn poll_until_sync(&mut self) -> Async<()> { ... }

    fn block_until_sync(&mut self, runtime: &mut Runtime) { ... }

    fn poll(&mut self) { ... }
}

Associated Types

type Specialization: NetworkSpecialization<Block> + SpecializationFactory

type Verifier: 'static + Verifier<Block>

type PeerData: Default

Loading content...
Required methods

fn from_config(config: &ProtocolConfig) -> Self

These two need to be implemented!

fn make_verifier(
&self,
client: PeersClient,
config: &ProtocolConfig,
peer_data: &Self::PeerData
) -> Self::Verifier

fn peer(&mut self, i: usize) -> &mut Peer<Self::PeerData, Self::Specialization>

Get reference to peer.

fn peers(&self) -> &Vec<Peer<Self::PeerData, Self::Specialization>>

fn mut_peers<F: FnOnce(&mut Vec<Peer<Self::PeerData, Self::Specialization>>)>(
&mut self,
closure: F
)

Loading content...
Provided methods

fn make_block_import(
&self,
client: PeersClient
) -> (BoxBlockImport<Block>, Option<BoxJustificationImport<Block>>, Option<BoxFinalityProofImport<Block>>, Option<BoxFinalityProofRequestBuilder<Block>>, Self::PeerData)

Get custom block import handle for fresh client, along with peer data.

fn make_finality_proof_provider(
&self,
_client: PeersClient
) -> Option<Arc<dyn FinalityProofProvider<Block>>>

Get finality proof provider (if supported).

fn default_config() -> ProtocolConfig

fn new(n: usize) -> Self

Create new test network with this many peers.

fn add_full_peer(&mut self, config: &ProtocolConfig)

fn add_full_peer_with_states(
&mut self,
config: &ProtocolConfig,
keep_blocks: Option<u32>
)

Add a full peer.

fn add_light_peer(&mut self, config: &ProtocolConfig)

Add a light peer.

fn poll_until_sync(&mut self) -> Async<()>

Polls the testnet until all nodes are in sync.

Must be executed in a task context.

fn block_until_sync(&mut self, runtime: &mut Runtime)

Blocks the current thread until we are sync'ed.

Calls poll_until_sync repeatidely with the runtime passed as parameter.

fn poll(&mut self)

Polls the testnet. Processes all the pending actions and returns NotReady.

Loading content...